stt_streaming_raw_packets/services/app/.ipynb_checkpoints/requirements-checkpoint.txt
asyncio==3.4.3
websockets==11.0.3
pydub==0.25.1
google-cloud-speech==2.19.0
